safari-600.7-branch/LayoutTests/js/dom/date-preserve-milliseconds-expected.txt

The following test checks if an existing milliseconds value gets preserved if a call to setHours(), setMinutes() or setSeconds() does not specify the milliseconds. See https://bugs.webkit.org/show_bug.cgi?id=3759

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.getMilliseconds() is 1
PASS d.getMilliseconds() is 1
PASS d.getMilliseconds() is 1
PASS d.getMilliseconds() is 1
PASS d.getMilliseconds() is 1
PASS d.getMilliseconds() is 1
PASS successfullyParsed is true

TEST COMPLETE
